+++ api/utils/errors/errors.unprocessable-entity.js
@@ -1,7 +1,7 @@
 const status = require('http-status')
 
 const defaultMessage = 'Could not process the instructions present'
-const ErrorGeneric = require('./erros.generic_error')
+const ErrorGeneric = require('./erros.generic-error')
 
 module.exports = class ErrorUnprocessableEntity extends ErrorGeneric {
   constructor(message) {
